// src/app/shared/interfaces/i-content-table.ts
export interface IContentTable {
  username: string;
  first_name: string;
  last_name: string;
  email: string;
  user_type: string;
}

export interface IContentUser {
  username: string;
  first_name: string;
  last_name: string;
  email: string;
  user_type: string;
  password: string;
  repeat_password: string
}